backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
pydantic==2.5.2
bcrypt==4.1.1
PyJWT==2.8.0
python-multipart==0.0.6
google-generativeai>=0.8.0
beautifulsoup4>=4.12.0
httpx>=0.27.0
requests>=2.31.0
